Remove ability to generate static methods from template 'keys'
As the Java compiler embeds the constants into the class using them, the
constants class itself is actually not loaded/needed at runtime.

By generating static methods, the constants class has to be loaded by
the JVM when they are used. If the constants class contains a lot of
constants, this is not desirable.